requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.0
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
git-filter-repo==2.47.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.67.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.1.1
grpcio==1.78.0
grpcio-status==1.71.2
grpcio-tools==1.78.0
gtts==2.2.4
gunicorn==21.2.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.27.0
huggingface-hub==0.23.4
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.6.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langgraph==0.1.19
langsmith==0.1.147
loguru==0.7.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy-extensions==1.1.0
numpy==1.26.4
openai==1.51.2
orjson==3.10.3
packaging==24.2
pandas==2.2.2
passlib==1.7.4
pillow==10.3.0
pluggy==1.6.0
portalocker==2.10.1
prompt-toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.3.4
pygments==2.19.2
pyjwt==2.8.0
pyparsing==3.3.2
pysqlite3==0.6.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.22
pytz==2026.1.post1
pywin32==311
pyyaml==6.0.3
qdrant-client==1.9.1
redis==5.0.4
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.4.2
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlalchemy==2.0.30
sqlparse==0.5.5
starlette==0.37.2
structlog==24.1.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
tqdm==4.67.3
transformers==4.41.2
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.29.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
whitenoise==6.7.0
win32-setctime==1.2.0
yarl==1.23.0
